extract unit test modules out to separate files
