+++ src/lib.rs
@@ -11,6 +11,8 @@ pub mod dfs_preorder_iterators;
 pub mod examples;
 pub mod leaves_iterators;
 pub mod prelude;
+#[cfg(test)]
+mod tests;
 mod tree_collection_iterators;
 mod tree_context;
 pub(crate) mod tree_iterators;
