+++ src/prelude.rs
@@ -4902,5097 +4902,3 @@ impl<'a, T> BorrowedIntoIteratorOfBinaryTrees<'a, T> for Iter<'a, T> where
 fn opt_to_opt<T>(opt: Option<T>) -> Option<T> {
     opt
 }
-
-#[cfg(test)]
-use streaming_iterator::StreamingIterator;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloc::vec;
-
-    #[cfg(test)]
-    extern crate std;
-    #[cfg(test)]
-    use std::collections::HashMap;
-    #[cfg(test)]
-    pub(crate) fn get_value_to_path_map() -> HashMap<usize, Vec<usize>> {
-        let mut result = HashMap::new();
-        result.insert(0, vec![]);
-        result.insert(1, vec![0]);
-        result.insert(2, vec![1]);
-        result.insert(3, vec![0, 0]);
-        result.insert(4, vec![0, 1]);
-        result.insert(5, vec![1, 0]);
-        result.insert(6, vec![1, 1]);
-        result.insert(7, vec![1, 1, 0]);
-        result.insert(8, vec![1, 1, 0, 0]);
-        result.insert(9, vec![1, 1, 0, 0, 0]);
-        result.insert(10, vec![1, 1, 0, 0, 0, 0]);
-        result
-    }
-
-    #[cfg(test)]
-    pub(crate) fn get_value_to_path_map_binary() -> HashMap<usize, Vec<usize>> {
-        let mut result = HashMap::new();
-        result.insert(0, vec![]);
-        result.insert(1, vec![0]);
-        result.insert(2, vec![1]);
-        result.insert(3, vec![0, 0]);
-        result.insert(4, vec![0, 1]);
-        result.insert(5, vec![1, 0]);
-        result.insert(6, vec![1, 1]);
-        result.insert(7, vec![1, 1, 0]);
-        result.insert(8, vec![1, 1, 0, 1]);
-        result.insert(9, vec![1, 1, 0, 1, 0]);
-        result.insert(10, vec![1, 1, 0, 1, 0, 1]);
-        result
-    }
-
-    #[cfg(test)]
-    mod dfs_preorder_tests {
-        use super::{
-            assert_len, create_binary_tree_for_testing, create_trees_for_testing,
-            get_expected_metadata_for_value, get_value_to_path_map, get_value_to_path_map_binary,
-        };
-        use crate::prelude::*;
-        use streaming_iterator::StreamingIterator;
-
-        pub(crate) fn get_expected_order_dfs_preorder() -> [usize; 11] {
-            [0, 1, 3, 4, 2, 5, 6, 7, 8, 9, 10]
-        }
-
-        #[test]
-        fn dfs_preorder_has_correct_order() {
-            let expected = get_expected_order_dfs_preorder();
-
-            for mut test_tree in create_trees_for_testing() {
-                for (i, value) in test_tree.dfs_preorder_iter().enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-                assert_len!(expected.len(), test_tree.dfs_preorder_iter());
-
-                for (i, value) in test_tree.dfs_preorder_iter_mut().enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-                assert_len!(expected.len(), test_tree.dfs_preorder_iter_mut());
-
-                for (i, value) in test_tree.clone().dfs_preorder().enumerate() {
-                    assert_eq!(expected[i], value);
-                }
-                assert_len!(expected.len(), test_tree.dfs_preorder());
-            }
-        }
-
-        #[test]
-        fn binary_dfs_preorder_has_correct_order() {
-            let expected = get_expected_order_dfs_preorder();
-
-            let mut test_tree = create_binary_tree_for_testing();
-            for (i, value) in test_tree.dfs_preorder_iter().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_preorder_iter());
-
-            for (i, value) in test_tree.dfs_preorder_iter_mut().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_preorder_iter_mut());
-
-            for (i, value) in test_tree.clone().dfs_preorder().enumerate() {
-                assert_eq!(expected[i], value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_preorder());
-        }
-
-        #[test]
-        fn dfs_preorder_attach_ancestors_works() {
-            let expected = get_expected_order_dfs_preorder();
-
-            for mut test_tree in create_trees_for_testing() {
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_preorder_iter().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_preorder_iter_mut().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_preorder().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-            }
-        }
-
-        #[test]
-        fn dfs_preorder_attach_context_works() {
-            let expected = get_expected_order_dfs_preorder();
-            let expected_paths = get_value_to_path_map();
-
-            for mut test_tree in create_trees_for_testing() {
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_preorder_iter().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        *value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value.ancestors[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(*value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_preorder_iter_mut().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        *value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(*value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_preorder().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-            }
-        }
-
-        #[test]
-        fn binary_dfs_preorder_attach_ancestors_works() {
-            let expected = get_expected_order_dfs_preorder();
-
-            let mut i = 0;
-            let test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.dfs_preorder_iter().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.dfs_preorder_iter_mut().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut iter_with_metadata = create_binary_tree_for_testing()
-                .dfs_preorder()
-                .attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-
-        #[test]
-        fn binary_dfs_preorder_attach_context_works() {
-            let expected = get_expected_order_dfs_preorder();
-            let expected_paths = get_value_to_path_map_binary();
-
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_preorder_iter().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(*value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_preorder_iter_mut().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(*value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_preorder().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                let expected =
-                    get_expected_metadata_for_value(value.ancestors()[value.ancestors().len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-    }
-
-    #[cfg(test)]
-    mod dfs_inorder_tests {
-        use super::{assert_len, create_binary_tree_for_testing, get_expected_metadata_for_value};
-        use crate::prelude::{tests::get_value_to_path_map_binary, *};
-        use streaming_iterator::StreamingIterator;
-
-        pub(crate) fn get_expected_order_dfs_inorder() -> [usize; 11] {
-            [3, 1, 4, 0, 5, 2, 7, 9, 10, 8, 6]
-        }
-
-        #[test]
-        fn dfs_inorder_has_correct_order() {
-            let expected = get_expected_order_dfs_inorder();
-
-            let mut test_tree = create_binary_tree_for_testing();
-            for (i, value) in test_tree.dfs_inorder_iter().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_inorder_iter());
-
-            for (i, value) in test_tree.dfs_inorder_iter_mut().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_inorder_iter_mut());
-
-            for (i, value) in test_tree.clone().dfs_inorder().enumerate() {
-                assert_eq!(expected[i], value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_inorder());
-        }
-
-        #[test]
-        fn dfs_inorder_attach_ancestors_works() {
-            let expected = get_expected_order_dfs_inorder();
-
-            let mut i = 0;
-            let test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.dfs_inorder_iter().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut iter_mut_with_metadata = test_tree.dfs_inorder_iter_mut().attach_ancestors();
-            while let Some(value) = iter_mut_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.dfs_inorder().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-
-        #[test]
-        fn binary_dfs_inorder_attach_context_works() {
-            let expected = get_expected_order_dfs_inorder();
-            let expected_paths = get_value_to_path_map_binary();
-
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_inorder_iter().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_inorder_iter_mut().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_inorder().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                let expected =
-                    get_expected_metadata_for_value(value.ancestors()[value.ancestors().len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-    }
-
-    mod dfs_postorder_tests {
-        use super::{
-            assert_len, create_binary_tree_for_testing, create_trees_for_testing,
-            get_expected_metadata_for_value, get_value_to_path_map,
-        };
-        use crate::prelude::{tests::get_value_to_path_map_binary, *};
-        use streaming_iterator::StreamingIterator;
-
-        pub(crate) fn get_expected_order_dfs_postorder() -> [usize; 11] {
-            [3, 4, 1, 5, 10, 9, 8, 7, 6, 2, 0]
-        }
-
-        #[test]
-        fn dfs_postorder_has_correct_order() {
-            let expected = get_expected_order_dfs_postorder();
-            for mut test_tree in create_trees_for_testing() {
-                for (i, value) in test_tree.dfs_postorder_iter().enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-                assert_len!(expected.len(), test_tree.dfs_postorder_iter());
-
-                for (i, value) in test_tree.dfs_postorder_iter_mut().enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-                assert_len!(expected.len(), test_tree.dfs_postorder_iter_mut());
-
-                for (i, value) in test_tree.clone().dfs_postorder().enumerate() {
-                    assert_eq!(expected[i], value);
-                }
-                assert_len!(expected.len(), test_tree.dfs_postorder());
-            }
-        }
-
-        #[test]
-        fn binary_dfs_postorder_has_correct_order() {
-            let expected = get_expected_order_dfs_postorder();
-            let mut test_tree = create_binary_tree_for_testing();
-
-            for (i, value) in test_tree.dfs_postorder_iter().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_postorder_iter());
-
-            for (i, value) in test_tree.dfs_postorder_iter_mut().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_postorder_iter_mut());
-
-            for (i, value) in test_tree.clone().dfs_postorder().enumerate() {
-                assert_eq!(expected[i], value);
-            }
-            assert_len!(expected.len(), test_tree.dfs_postorder());
-        }
-
-        #[test]
-        fn dfs_postorder_attach_ancestors_works() {
-            let expected = get_expected_order_dfs_postorder();
-
-            for mut test_tree in create_trees_for_testing() {
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_postorder_iter().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_postorder_iter_mut().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_postorder().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-            }
-        }
-
-        #[test]
-        fn binary_dfs_postorder_attach_context_works() {
-            let expected = get_expected_order_dfs_postorder();
-            let expected_paths = get_value_to_path_map_binary();
-
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_postorder_iter().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_postorder_iter_mut().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.dfs_postorder().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                let expected =
-                    get_expected_metadata_for_value(value.ancestors()[value.ancestors().len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-
-        #[test]
-        fn dfs_postorder_attach_context_works() {
-            let expected = get_expected_order_dfs_postorder();
-            let expected_paths = get_value_to_path_map();
-
-            for mut test_tree in create_trees_for_testing() {
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_postorder_iter().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        *value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(*value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_postorder_iter_mut().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        *value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(*value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.dfs_postorder().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-            }
-        }
-
-        #[test]
-        fn binary_dfs_postorder_attach_ancestors_works() {
-            let expected = get_expected_order_dfs_postorder();
-
-            let mut i = 0;
-            let test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.dfs_postorder_iter().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.dfs_postorder_iter_mut().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut iter_with_metadata = create_binary_tree_for_testing()
-                .dfs_postorder()
-                .attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-    }
-
-    mod bfs_tests {
-        use super::{
-            assert_len, create_binary_tree_for_testing, create_trees_for_testing,
-            get_expected_metadata_for_value, get_value_to_path_map, get_value_to_path_map_binary,
-        };
-        use crate::prelude::*;
-        use streaming_iterator::StreamingIterator;
-
-        pub(crate) fn get_expected_order_bfs() -> [usize; 11] {
-            [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-        }
-
-        #[test]
-        fn bfs_has_correct_order() {
-            let expected = get_expected_order_bfs();
-            for mut test_tree in create_trees_for_testing() {
-                for (i, value) in test_tree.bfs_iter().enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-                assert_len!(expected.len(), test_tree.bfs_iter());
-
-                for (i, value) in test_tree.bfs_iter_mut().enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-                assert_len!(expected.len(), test_tree.bfs_iter_mut());
-
-                for (i, value) in test_tree.clone().bfs().enumerate() {
-                    assert_eq!(expected[i], value);
-                }
-                assert_len!(expected.len(), test_tree.bfs());
-            }
-        }
-
-        #[test]
-        fn bfs_attach_ancestors_works() {
-            let expected = get_expected_order_bfs();
-
-            for mut test_tree in create_trees_for_testing() {
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.bfs_iter().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.bfs_iter_mut().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.bfs().attach_ancestors();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], value[value.len() - 1]);
-                    let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], value[j]);
-                    }
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-            }
-        }
-
-        #[test]
-        fn bfs_attach_context_works() {
-            let expected = get_expected_order_bfs();
-            let expected_paths = get_value_to_path_map();
-
-            for mut test_tree in create_trees_for_testing() {
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.bfs_iter().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        *value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.bfs_iter_mut().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        *value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], *value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-                drop(iter_with_metadata);
-
-                let mut i = 0;
-                let mut iter_with_metadata = test_tree.bfs().attach_context();
-                while let Some(value) = iter_with_metadata.next() {
-                    assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                    let expected = get_expected_metadata_for_value(
-                        value.ancestors()[value.ancestors().len() - 1],
-                    );
-                    for j in 0..expected.len() {
-                        assert_eq!(expected[j], value.ancestors()[j]);
-                    }
-                    assert_eq!(
-                        *expected_paths
-                            .get(value.ancestors().last().unwrap())
-                            .unwrap(),
-                        value.path
-                    );
-                    i += 1;
-                }
-                assert_eq!(expected.len(), i);
-            }
-        }
-
-        #[test]
-        fn binary_bfs_has_correct_order() {
-            let expected = get_expected_order_bfs();
-            let mut test_tree = create_binary_tree_for_testing();
-
-            for (i, value) in test_tree.bfs_iter().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.bfs_iter());
-
-            for (i, value) in test_tree.bfs_iter_mut().enumerate() {
-                assert_eq!(expected[i], *value);
-            }
-            assert_len!(expected.len(), test_tree.bfs_iter_mut());
-
-            for (i, value) in test_tree.clone().bfs().enumerate() {
-                assert_eq!(expected[i], value);
-            }
-            assert_len!(expected.len(), test_tree.bfs());
-        }
-
-        #[test]
-        fn binary_bfs_attach_context_works() {
-            let expected = get_expected_order_bfs();
-            let expected_paths = get_value_to_path_map_binary();
-
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.bfs_iter().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.bfs_iter_mut().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value.ancestors()[value.ancestors().len() - 1]);
-                let expected = get_expected_metadata_for_value(
-                    *value.ancestors()[value.ancestors().len() - 1],
-                );
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-            drop(iter_with_metadata);
-
-            let mut i = 0;
-            let mut iter_with_metadata = test_tree.bfs().attach_context();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value.ancestors()[value.ancestors().len() - 1]);
-                let expected =
-                    get_expected_metadata_for_value(value.ancestors()[value.ancestors().len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value.ancestors()[j]);
-                }
-                assert_eq!(
-                    *expected_paths
-                        .get(value.ancestors().last().unwrap())
-                        .unwrap(),
-                    value.path
-                );
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-
-        #[test]
-        fn binary_bfs_attach_ancestors_works() {
-            let expected = get_expected_order_bfs();
-
-            let mut i = 0;
-            let test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.bfs_iter().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut test_tree = create_binary_tree_for_testing();
-            let mut iter_with_metadata = test_tree.bfs_iter_mut().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], *value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(*value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], *value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-
-            let mut i = 0;
-            let mut iter_with_metadata = create_binary_tree_for_testing().bfs().attach_ancestors();
-            while let Some(value) = iter_with_metadata.next() {
-                assert_eq!(expected[i], value[value.len() - 1]);
-                let expected = get_expected_metadata_for_value(value[value.len() - 1]);
-                for j in 0..expected.len() {
-                    assert_eq!(expected[j], value[j]);
-                }
-                i += 1;
-            }
-            assert_eq!(expected.len(), i);
-        }
-    }
-
-    #[cfg(test)]
-    mod ancestors_leaves_tests {
-        use alloc::{string::ToString, vec};
-
-        use super::{assert_len, create_binary_tree_for_testing, create_trees_for_testing};
-        use crate::prelude::*;
-
-        fn get_expected_order_leaves() -> [Vec<usize>; 4] {
-            [
-                vec![0, 1, 3],
-                vec![0, 1, 4],
-                vec![0, 2, 5],
-                vec![0, 2, 6, 7, 8, 9, 10],
-            ]
-        }
-
-        #[test]
-        fn leaves_has_correct_order() {
-            let expected = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                for mut borrowed_iter in get_borrowed_leaves_iters(&test_tree) {
-                    let mut i = 0;
-                    while let Some(value) = borrowed_iter.next() {
-                        assert!(expected[i].iter().eq(value.iter().map(|val| *val)));
-                        i += 1;
-                    }
-                }
-
-                let mut i = 0;
-                for borrowed_iter in get_borrowed_leaves_iters(&test_tree) {
-                    assert_len!(
-                        expected.len(),
-                        borrowed_iter,
-                        alloc::format!("Failure at index {}", i.to_string())
-                    );
-                    i += 1;
-                }
-
-                for mut mut_borrowed_iter in get_mut_borrowed_leaves_iters(&mut test_tree) {
-                    let mut i = 0;
-                    while let Some(value) = mut_borrowed_iter.next() {
-                        assert!(expected[i].iter().eq(value.iter().map(|val| &**val)));
-                        i += 1;
-                    }
-                }
-
-                for mut_borrowed_iter in get_mut_borrowed_leaves_iters(&mut test_tree) {
-                    assert_len!(expected.len(), mut_borrowed_iter);
-                }
-
-                for mut owned_iter in get_owned_leaves_iters(test_tree.clone()) {
-                    let mut i = 0;
-                    while let Some(value) = owned_iter.next() {
-                        assert!(expected[i].iter().eq(value.iter()));
-                        i += 1;
-                    }
-                }
-
-                for owned_iter in get_owned_leaves_iters(test_tree) {
-                    assert_len!(expected.len(), owned_iter);
-                }
-            }
-        }
-
-        fn get_borrowed_leaves_iters<T>(
-            test_tree: &Tree<T>,
-        ) -> impl Iterator<Item = Box<dyn StreamingIterator<Item = [&T]> + '_>> + '_ {
-            [
-                Box::new(test_tree.dfs_preorder_iter().attach_ancestors().leaves())
-                    as Box<dyn StreamingIterator<Item = [&T]>>,
-                Box::new(test_tree.dfs_postorder_iter().attach_ancestors().leaves()),
-                Box::new(test_tree.bfs_iter().attach_ancestors().leaves()),
-            ]
-            .into_iter()
-        }
-
-        fn get_mut_borrowed_leaves_iters<T>(
-            test_tree: &mut Tree<T>,
-        ) -> impl Iterator<Item = Box<dyn StreamingIterator<Item = [&mut T]> + '_>> + '_ {
-            // Rust doesn't like this, but we know that only 1 iterator will be accessed at a time
-            // and no reallocations will be done as we are doing a readonly test,
-            // so we are still within the "safe" rust system with only 1 active mutable reference.
-            // This also makes the test much nicer to write.
-            unsafe {
-                [
-                    Box::new(
-                        (*(test_tree as *mut Tree<T>))
-                            .dfs_preorder_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ) as Box<dyn StreamingIterator<Item = [&mut T]>>,
-                    Box::new(
-                        (*(test_tree as *mut Tree<T>))
-                            .dfs_postorder_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ),
-                    Box::new(
-                        (*(test_tree as *mut Tree<T>))
-                            .bfs_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ),
-                ]
-                .into_iter()
-            }
-        }
-
-        fn get_owned_leaves_iters<T: Clone + 'static>(
-            test_tree: Tree<T>,
-        ) -> [Box<dyn StreamingIterator<Item = [T]>>; 3] {
-            [
-                Box::new(test_tree.clone().dfs_preorder().attach_ancestors().leaves())
-                    as Box<dyn StreamingIterator<Item = [T]>>,
-                Box::new(
-                    test_tree
-                        .clone()
-                        .dfs_postorder()
-                        .attach_ancestors()
-                        .leaves(),
-                ),
-                Box::new(test_tree.clone().bfs().attach_ancestors().leaves()),
-            ]
-        }
-
-        #[test]
-        fn binary_leaves_has_correct_order() {
-            let expected = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-
-            for mut borrowed_iter in get_borrowed_leaves_binary_iters(&test_tree) {
-                let mut i = 0;
-                while let Some(value) = borrowed_iter.next() {
-                    assert!(expected[i].iter().eq(value.iter().map(|val| *val)));
-                    i += 1;
-                }
-            }
-
-            let mut results = Vec::new();
-            for borrowed_iter in get_borrowed_leaves_binary_iters(&test_tree) {
-                let mut count = 0;
-                borrowed_iter.for_each(|_| count += 1);
-                results.push(count);
-            }
-
-            let mut i = 0;
-            for borrowed_iter in get_borrowed_leaves_binary_iters(&test_tree) {
-                assert_len!(
-                    expected.len(),
-                    borrowed_iter,
-                    alloc::format!("Failure at index {}", i.to_string())
-                );
-                i += 1;
-            }
-
-            for mut mut_borrowed_iter in get_mut_borrowed_leaves_binary_iters(&mut test_tree) {
-                let mut i = 0;
-                while let Some(value) = mut_borrowed_iter.next() {
-                    assert!(expected[i].iter().eq(value.iter().map(|val| &**val)));
-                    i += 1;
-                }
-            }
-
-            for mut_borrowed_iter in get_mut_borrowed_leaves_binary_iters(&mut test_tree) {
-                assert_len!(expected.len(), mut_borrowed_iter);
-            }
-
-            for mut owned_iter in get_owned_leaves_binary_iters(test_tree.clone()) {
-                let mut i = 0;
-                while let Some(value) = owned_iter.next() {
-                    assert!(expected[i].iter().eq(value.iter()));
-                    i += 1;
-                }
-            }
-
-            for owned_iter in get_owned_leaves_binary_iters(test_tree) {
-                assert_len!(expected.len(), owned_iter);
-            }
-        }
-
-        fn get_borrowed_leaves_binary_iters<T>(
-            test_tree: &BinaryTree<T>,
-        ) -> [Box<dyn StreamingIterator<Item = [&T]> + '_>; 4] {
-            [
-                Box::new(test_tree.dfs_preorder_iter().attach_ancestors().leaves())
-                    as Box<dyn StreamingIterator<Item = [&T]>>,
-                Box::new(test_tree.dfs_inorder_iter().attach_ancestors().leaves()),
-                Box::new(test_tree.dfs_postorder_iter().attach_ancestors().leaves()),
-                Box::new(test_tree.bfs_iter().attach_ancestors().leaves()),
-            ]
-        }
-
-        fn get_mut_borrowed_leaves_binary_iters<T>(
-            test_tree: &mut BinaryTree<T>,
-        ) -> impl Iterator<Item = Box<dyn StreamingIterator<Item = [&mut T]> + '_>> {
-            unsafe {
-                [
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .dfs_preorder_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ) as Box<dyn StreamingIterator<Item = [&mut T]>>,
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .dfs_inorder_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ),
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .dfs_postorder_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ),
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .bfs_iter_mut()
-                            .attach_ancestors()
-                            .leaves(),
-                    ),
-                ]
-                .into_iter()
-            }
-        }
-
-        fn get_owned_leaves_binary_iters<T: Clone + 'static>(
-            test_tree: BinaryTree<T>,
-        ) -> [Box<dyn StreamingIterator<Item = [T]>>; 4] {
-            [
-                Box::new(test_tree.clone().dfs_preorder().attach_ancestors().leaves())
-                    as Box<dyn StreamingIterator<Item = [T]>>,
-                Box::new(test_tree.clone().dfs_inorder().attach_ancestors().leaves()),
-                Box::new(
-                    test_tree
-                        .clone()
-                        .dfs_postorder()
-                        .attach_ancestors()
-                        .leaves(),
-                ),
-                Box::new(test_tree.clone().bfs().attach_ancestors().leaves()),
-            ]
-        }
-
-        #[test]
-        fn dfs_preorder_transformation_can_happen_mid_traversal() {
-            let expected_dfs_preorder =
-                super::dfs_preorder_tests::get_expected_order_dfs_preorder();
-            let expected_leaves = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                // interrupt traversal at all points.
-                for _ in 0..expected_dfs_preorder.len() {
-                    let mut preorder_iter = test_tree.dfs_preorder_iter().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = preorder_iter.next() {
-                        if *value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut preorder_iter_leaves = preorder_iter.leaves();
-                    let mut i = 0;
-                    while let Some(value) = preorder_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                        i += 1;
-                    }
-                    drop(preorder_iter_leaves);
-
-                    let mut preorder_iter_mut =
-                        test_tree.dfs_preorder_iter_mut().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = preorder_iter_mut.next() {
-                        if *value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut preorder_iter_leaves = preorder_iter_mut.leaves();
-                    let mut i = 0;
-                    while let Some(value) = preorder_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                        i += 1;
-                    }
-                    drop(preorder_iter_leaves);
-
-                    let mut preorder = test_tree.clone().dfs_preorder().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = preorder.next() {
-                        if value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut preorder_iter_leaves = preorder.leaves();
-                    let mut i = 0;
-                    while let Some(value) = preorder_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter()));
-                        i += 1;
-                    }
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_postorder_transformation_can_happen_mid_traversal() {
-            let expected_dfs_postorder =
-                super::dfs_postorder_tests::get_expected_order_dfs_postorder();
-            let expected_leaves = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                // interrupt traversal at all points.
-                for _ in 0..expected_dfs_postorder.len() {
-                    let mut postorder_iter = test_tree.dfs_postorder_iter().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = postorder_iter.next() {
-                        // dont index outside the array!
-                        if num_leaves_seen == expected_leaves.len() {
-                            continue;
-                        }
-                        if *value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut postorder_iter_leaves = postorder_iter.leaves();
-                    let mut i = 0;
-                    while let Some(value) = postorder_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                        i += 1;
-                    }
-                    drop(postorder_iter_leaves);
-
-                    let mut postorder_iter_mut =
-                        test_tree.dfs_postorder_iter_mut().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = postorder_iter_mut.next() {
-                        // dont index outside the array!
-                        if num_leaves_seen == expected_leaves.len() {
-                            continue;
-                        }
-                        if *value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut postorder_iter_leaves = postorder_iter_mut.leaves();
-                    let mut i = 0;
-                    while let Some(value) = postorder_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                        i += 1;
-                    }
-                    drop(postorder_iter_leaves);
-
-                    let mut postorder = test_tree.clone().dfs_postorder().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = postorder.next() {
-                        // dont index outside the array!
-                        if num_leaves_seen == expected_leaves.len() {
-                            continue;
-                        }
-                        if value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut postorder_iter_leaves = postorder.leaves();
-                    let mut i = 0;
-                    while let Some(value) = postorder_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter()));
-                        i += 1;
-                    }
-                }
-            }
-        }
-
-        #[test]
-        fn bfs_transformation_can_happen_mid_traversal() {
-            let expected_bfs = super::bfs_tests::get_expected_order_bfs();
-            let expected_leaves = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                // interrupt traversal at all points.
-                for _ in 0..expected_bfs.len() {
-                    let mut bfs_iter = test_tree.bfs_iter().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = bfs_iter.next() {
-                        if *value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut bfs_iter_leaves = bfs_iter.leaves();
-                    let mut i = 0;
-                    while let Some(value) = bfs_iter_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                        i += 1;
-                    }
-                    drop(bfs_iter_leaves);
-
-                    let mut bfs_iter_mut = test_tree.bfs_iter_mut().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = bfs_iter_mut.next() {
-                        if *value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut bfs_iter_mut_leaves = bfs_iter_mut.leaves();
-                    let mut i = 0;
-                    while let Some(value) = bfs_iter_mut_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                        i += 1;
-                    }
-                    drop(bfs_iter_mut_leaves);
-
-                    let mut bfs = test_tree.clone().bfs().attach_ancestors();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = bfs.next() {
-                        if value[value.len() - 1]
-                            == expected_leaves[num_leaves_seen]
-                                [expected_leaves[num_leaves_seen].len() - 1]
-                        {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let mut bfs_leaves = bfs.leaves();
-                    let mut i = 0;
-                    while let Some(value) = bfs_leaves.next() {
-                        assert!(expected_leaves[i].iter().eq(value.iter()));
-                        i += 1;
-                    }
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_preorder_binary_transformation_can_happen_mid_traversal() {
-            let expected_dfs_preorder =
-                super::dfs_preorder_tests::get_expected_order_dfs_preorder();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_dfs_preorder.len() {
-                let mut preorder_iter = test_tree.dfs_preorder_iter().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = preorder_iter.next() {
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut preorder_iter_leaves = preorder_iter.leaves();
-                let mut i = 0;
-                while let Some(value) = preorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                    i += 1;
-                }
-                drop(preorder_iter_leaves);
-
-                let mut preorder_iter_mut = test_tree.dfs_preorder_iter_mut().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = preorder_iter_mut.next() {
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut preorder_iter_leaves = preorder_iter_mut.leaves();
-                let mut i = 0;
-                while let Some(value) = preorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                    i += 1;
-                }
-                drop(preorder_iter_leaves);
-
-                let mut preorder = test_tree.clone().dfs_preorder().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = preorder.next() {
-                    if value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut preorder_iter_leaves = preorder.leaves();
-                let mut i = 0;
-                while let Some(value) = preorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter()));
-                    i += 1;
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_inorder_binary_transformation_can_happen_mid_traversal() {
-            let expected_dfs_inorder = super::dfs_inorder_tests::get_expected_order_dfs_inorder();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_dfs_inorder.len() {
-                let mut inorder_iter = test_tree.dfs_inorder_iter().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = inorder_iter.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut inorder_iter_leaves = inorder_iter.leaves();
-                let mut i = 0;
-                while let Some(value) = inorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                    i += 1;
-                }
-                drop(inorder_iter_leaves);
-
-                let mut inorder_iter_mut = test_tree.dfs_inorder_iter_mut().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = inorder_iter_mut.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut inorder_iter_leaves = inorder_iter_mut.leaves();
-                let mut i = 0;
-                while let Some(value) = inorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                    i += 1;
-                }
-                drop(inorder_iter_leaves);
-
-                let mut inorder = test_tree.clone().dfs_inorder().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = inorder.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut inorder_iter_leaves = inorder.leaves();
-                let mut i = 0;
-                while let Some(value) = inorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter()));
-                    i += 1;
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_postorder_binary_transformation_can_happen_mid_traversal() {
-            let expected_dfs_postorder =
-                super::dfs_postorder_tests::get_expected_order_dfs_postorder();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_dfs_postorder.len() {
-                let mut postorder_iter = test_tree.dfs_postorder_iter().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = postorder_iter.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut postorder_iter_leaves = postorder_iter.leaves();
-                let mut i = 0;
-                while let Some(value) = postorder_iter_leaves.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                    i += 1;
-                }
-                drop(postorder_iter_leaves);
-
-                let mut postorder_iter_mut = test_tree.dfs_postorder_iter_mut().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = postorder_iter_mut.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut postorder_iter_leaves = postorder_iter_mut.leaves();
-                let mut i = 0;
-                while let Some(value) = postorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                    i += 1;
-                }
-                drop(postorder_iter_leaves);
-
-                let mut postorder = test_tree.clone().dfs_postorder().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = postorder.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut postorder_iter_leaves = postorder.leaves();
-                let mut i = 0;
-                while let Some(value) = postorder_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter()));
-                    i += 1;
-                }
-            }
-        }
-
-        #[test]
-        fn bfs_binary_transformation_can_happen_mid_traversal() {
-            let expected_bfs = super::bfs_tests::get_expected_order_bfs();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_bfs.len() {
-                let mut bfs_iter = test_tree.bfs_iter().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = bfs_iter.next() {
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut bfs_iter_leaves = bfs_iter.leaves();
-                let mut i = 0;
-                while let Some(value) = bfs_iter_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| *val)));
-                    i += 1;
-                }
-                drop(bfs_iter_leaves);
-
-                let mut bfs_iter_mut = test_tree.bfs_iter_mut().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = bfs_iter_mut.next() {
-                    if *value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut bfs_iter_mut_leaves = bfs_iter_mut.leaves();
-                let mut i = 0;
-                while let Some(value) = bfs_iter_mut_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter().map(|val| &**val)));
-                    i += 1;
-                }
-                drop(bfs_iter_mut_leaves);
-
-                let mut bfs = test_tree.clone().bfs().attach_ancestors();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = bfs.next() {
-                    if value[value.len() - 1]
-                        == expected_leaves[num_leaves_seen]
-                            [expected_leaves[num_leaves_seen].len() - 1]
-                    {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let mut bfs_leaves = bfs.leaves();
-                let mut i = 0;
-                while let Some(value) = bfs_leaves.next() {
-                    assert!(expected_leaves[i].iter().eq(value.iter()));
-                    i += 1;
-                }
-            }
-        }
-    }
-
-    #[cfg(test)]
-    mod leaves_tests {
-        use super::{assert_len, create_binary_tree_for_testing, create_trees_for_testing};
-        use crate::prelude::*;
-
-        fn get_expected_order_leaves() -> [usize; 4] {
-            [3, 4, 5, 10]
-        }
-
-        #[test]
-        fn leaves_has_correct_order() {
-            let expected = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                for borrowed_iter in get_borrowed_leaves_iters(&test_tree) {
-                    for (i, value) in borrowed_iter.enumerate() {
-                        assert_eq!(expected[i], *value);
-                    }
-                }
-
-                for borrowed_iter in get_borrowed_leaves_iters(&test_tree) {
-                    assert_len!(expected.len(), borrowed_iter);
-                }
-
-                for mut_borrowed_iter in get_mut_borrowed_leaves_iters(&mut test_tree) {
-                    for (i, value) in mut_borrowed_iter.enumerate() {
-                        assert_eq!(expected[i], *value);
-                    }
-                }
-
-                for mut_borrowed_iter in get_mut_borrowed_leaves_iters(&mut test_tree) {
-                    assert_len!(expected.len(), mut_borrowed_iter);
-                }
-
-                for owned_iter in get_owned_leaves_iters(test_tree.clone()) {
-                    for (i, value) in owned_iter.enumerate() {
-                        assert_eq!(expected[i], value);
-                    }
-                }
-
-                for owned_iter in get_owned_leaves_iters(test_tree) {
-                    assert_len!(expected.len(), owned_iter);
-                }
-            }
-        }
-
-        fn get_borrowed_leaves_iters<T>(
-            test_tree: &Tree<T>,
-        ) -> impl Iterator<Item = Box<dyn Iterator<Item = &T> + '_>> + '_ {
-            [
-                Box::new(test_tree.dfs_preorder_iter().leaves()) as Box<dyn Iterator<Item = &T>>,
-                Box::new(test_tree.dfs_postorder_iter().leaves()),
-                Box::new(test_tree.bfs_iter().leaves()),
-            ]
-            .into_iter()
-        }
-
-        fn get_mut_borrowed_leaves_iters<T>(
-            test_tree: &mut Tree<T>,
-        ) -> impl Iterator<Item = Box<dyn Iterator<Item = &mut T> + '_>> + '_ {
-            // Rust doesn't like this, but we know that only 1 iterator will be accessed at a time
-            // and no reallocations will be done as we are doing a readonly test,
-            // so we are still within the "safe" rust system with only 1 active mutable reference.
-            // This also makes the test much nicer to write.
-            unsafe {
-                [
-                    Box::new(
-                        (*(test_tree as *mut Tree<T>))
-                            .dfs_preorder_iter_mut()
-                            .leaves(),
-                    ) as Box<dyn Iterator<Item = &mut T>>,
-                    Box::new(
-                        (*(test_tree as *mut Tree<T>))
-                            .dfs_postorder_iter_mut()
-                            .leaves(),
-                    ),
-                    Box::new((*(test_tree as *mut Tree<T>)).bfs_iter_mut().leaves()),
-                ]
-                .into_iter()
-            }
-        }
-
-        fn get_owned_leaves_iters<T: Clone + 'static>(
-            test_tree: Tree<T>,
-        ) -> [Box<dyn Iterator<Item = T>>; 3] {
-            [
-                Box::new(test_tree.clone().dfs_preorder().leaves()) as Box<dyn Iterator<Item = T>>,
-                Box::new(test_tree.clone().dfs_postorder().leaves()),
-                Box::new(test_tree.clone().bfs().leaves()),
-            ]
-        }
-
-        #[test]
-        fn binary_leaves_has_correct_order() {
-            let expected = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-
-            for borrowed_iter in get_borrowed_leaves_binary_iters(&test_tree) {
-                for (i, value) in borrowed_iter.enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-            }
-
-            for borrowed_iter in get_borrowed_leaves_binary_iters(&test_tree) {
-                assert_len!(expected.len(), borrowed_iter);
-            }
-
-            for mut_borrowed_iter in get_mut_borrowed_leaves_binary_iters(&mut test_tree) {
-                for (i, value) in mut_borrowed_iter.enumerate() {
-                    assert_eq!(expected[i], *value);
-                }
-            }
-
-            for mut_borrowed_iter in get_mut_borrowed_leaves_binary_iters(&mut test_tree) {
-                assert_len!(expected.len(), mut_borrowed_iter);
-            }
-
-            for owned_iter in get_owned_leaves_binary_iters(test_tree.clone()) {
-                for (i, value) in owned_iter.enumerate() {
-                    assert_eq!(expected[i], value);
-                }
-            }
-
-            for owned_iter in get_owned_leaves_binary_iters(test_tree) {
-                assert_len!(expected.len(), owned_iter);
-            }
-        }
-
-        fn get_borrowed_leaves_binary_iters<T>(
-            test_tree: &BinaryTree<T>,
-        ) -> [Box<dyn Iterator<Item = &T> + '_>; 4] {
-            [
-                Box::new(test_tree.dfs_preorder_iter().leaves()) as Box<dyn Iterator<Item = &T>>,
-                Box::new(test_tree.dfs_inorder_iter().leaves()),
-                Box::new(test_tree.dfs_postorder_iter().leaves()),
-                Box::new(test_tree.bfs_iter().leaves()),
-            ]
-        }
-
-        fn get_mut_borrowed_leaves_binary_iters<T>(
-            test_tree: &mut BinaryTree<T>,
-        ) -> impl Iterator<Item = Box<dyn Iterator<Item = &mut T> + '_>> {
-            unsafe {
-                [
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .dfs_preorder_iter_mut()
-                            .leaves(),
-                    ) as Box<dyn Iterator<Item = &mut T>>,
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .dfs_inorder_iter_mut()
-                            .leaves(),
-                    ),
-                    Box::new(
-                        (*(test_tree as *mut BinaryTree<T>))
-                            .dfs_postorder_iter_mut()
-                            .leaves(),
-                    ),
-                    Box::new((*(test_tree as *mut BinaryTree<T>)).bfs_iter_mut().leaves()),
-                ]
-                .into_iter()
-            }
-        }
-
-        fn get_owned_leaves_binary_iters<T: Clone + 'static>(
-            test_tree: BinaryTree<T>,
-        ) -> [Box<dyn Iterator<Item = T>>; 4] {
-            [
-                Box::new(test_tree.clone().dfs_preorder().leaves()) as Box<dyn Iterator<Item = T>>,
-                Box::new(test_tree.clone().dfs_inorder().leaves()),
-                Box::new(test_tree.clone().dfs_postorder().leaves()),
-                Box::new(test_tree.clone().bfs().leaves()),
-            ]
-        }
-
-        #[test]
-        fn dfs_preorder_transformation_can_happen_mid_traversal() {
-            let expected_dfs_preorder =
-                super::dfs_preorder_tests::get_expected_order_dfs_preorder();
-            let expected_leaves = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                // interrupt traversal at all points.
-                for _ in 0..expected_dfs_preorder.len() {
-                    let mut preorder_iter = test_tree.dfs_preorder_iter();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = preorder_iter.next() {
-                        if *value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let preorder_iter_leaves = preorder_iter.leaves();
-                    for (i, value) in preorder_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                    }
-
-                    let mut preorder_iter_mut = test_tree.dfs_preorder_iter_mut();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = preorder_iter_mut.next() {
-                        if *value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let preorder_iter_leaves = preorder_iter_mut.leaves();
-                    for (i, value) in preorder_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                    }
-
-                    let mut preorder = test_tree.clone().dfs_preorder();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = preorder.next() {
-                        if value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let preorder_iter_leaves = preorder.leaves();
-                    for (i, value) in preorder_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                    }
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_postorder_transformation_can_happen_mid_traversal() {
-            let expected_dfs_postorder =
-                super::dfs_postorder_tests::get_expected_order_dfs_postorder();
-            let expected_leaves = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                // interrupt traversal at all points.
-                for _ in 0..expected_dfs_postorder.len() {
-                    let mut postorder_iter = test_tree.dfs_postorder_iter();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = postorder_iter.next() {
-                        // dont index outside the array!
-                        if num_leaves_seen == expected_leaves.len() {
-                            continue;
-                        }
-                        if *value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let postorder_iter_leaves = postorder_iter.leaves();
-                    for (i, value) in postorder_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                    }
-
-                    let mut postorder_iter_mut = test_tree.dfs_postorder_iter_mut();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = postorder_iter_mut.next() {
-                        // dont index outside the array!
-                        if num_leaves_seen == expected_leaves.len() {
-                            continue;
-                        }
-                        if *value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let postorder_iter_leaves = postorder_iter_mut.leaves();
-                    for (i, value) in postorder_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                    }
-
-                    let mut postorder = test_tree.clone().dfs_postorder();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = postorder.next() {
-                        // dont index outside the array!
-                        if num_leaves_seen == expected_leaves.len() {
-                            continue;
-                        }
-                        if value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let postorder_iter_leaves = postorder.leaves();
-                    for (i, value) in postorder_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                    }
-                }
-            }
-        }
-
-        #[test]
-        fn bfs_transformation_can_happen_mid_traversal() {
-            let expected_bfs = super::bfs_tests::get_expected_order_bfs();
-            let expected_leaves = get_expected_order_leaves();
-            for mut test_tree in create_trees_for_testing() {
-                // interrupt traversal at all points.
-                for _ in 0..expected_bfs.len() {
-                    let mut bfs_iter = test_tree.bfs_iter();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = bfs_iter.next() {
-                        if *value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let bfs_iter_leaves = bfs_iter.leaves();
-                    for (i, value) in bfs_iter_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                    }
-
-                    let mut bfs_iter_mut = test_tree.bfs_iter_mut();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = bfs_iter_mut.next() {
-                        if *value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let bfs_iter_mut_leaves = bfs_iter_mut.leaves();
-                    for (i, value) in bfs_iter_mut_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                    }
-
-                    let mut bfs = test_tree.clone().bfs();
-                    let mut num_leaves_seen = 0;
-                    while let Some(value) = bfs.next() {
-                        if value == expected_leaves[num_leaves_seen] {
-                            num_leaves_seen += 1;
-                        }
-                    }
-
-                    let bfs_leaves = bfs.leaves();
-                    for (i, value) in bfs_leaves.enumerate() {
-                        assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                    }
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_preorder_binary_transformation_can_happen_mid_traversal() {
-            let expected_dfs_preorder =
-                super::dfs_preorder_tests::get_expected_order_dfs_preorder();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_dfs_preorder.len() {
-                let mut preorder_iter = test_tree.dfs_preorder_iter();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = preorder_iter.next() {
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let preorder_iter_leaves = preorder_iter.leaves();
-                for (i, value) in preorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut preorder_iter_mut = test_tree.dfs_preorder_iter_mut();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = preorder_iter_mut.next() {
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let preorder_iter_leaves = preorder_iter_mut.leaves();
-                for (i, value) in preorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut preorder = test_tree.clone().dfs_preorder();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = preorder.next() {
-                    if value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let preorder_iter_leaves = preorder.leaves();
-                for (i, value) in preorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_inorder_binary_transformation_can_happen_mid_traversal() {
-            let expected_dfs_inorder = super::dfs_inorder_tests::get_expected_order_dfs_inorder();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_dfs_inorder.len() {
-                let mut inorder_iter = test_tree.dfs_inorder_iter();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = inorder_iter.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let inorder_iter_leaves = inorder_iter.leaves();
-                for (i, value) in inorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut inorder_iter_mut = test_tree.dfs_inorder_iter_mut();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = inorder_iter_mut.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let inorder_iter_leaves = inorder_iter_mut.leaves();
-                for (i, value) in inorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut inorder = test_tree.clone().dfs_inorder();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = inorder.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let inorder_iter_leaves = inorder.leaves();
-                for (i, value) in inorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                }
-            }
-        }
-
-        #[test]
-        fn dfs_postorder_binary_transformation_can_happen_mid_traversal() {
-            let expected_dfs_postorder =
-                super::dfs_postorder_tests::get_expected_order_dfs_postorder();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_dfs_postorder.len() {
-                let mut postorder_iter = test_tree.dfs_postorder_iter();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = postorder_iter.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let postorder_iter_leaves = postorder_iter.leaves();
-                for (i, value) in postorder_iter_leaves.enumerate() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut postorder_iter_mut = test_tree.dfs_postorder_iter_mut();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = postorder_iter_mut.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let postorder_iter_leaves = postorder_iter_mut.leaves();
-                for (i, value) in postorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut postorder = test_tree.clone().dfs_postorder();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = postorder.next() {
-                    // dont index outside the array!
-                    if num_leaves_seen == expected_leaves.len() {
-                        continue;
-                    }
-                    if value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let postorder_iter_leaves = postorder.leaves();
-                for (i, value) in postorder_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                }
-            }
-        }
-
-        #[test]
-        fn bfs_binary_transformation_can_happen_mid_traversal() {
-            let expected_bfs = super::bfs_tests::get_expected_order_bfs();
-            let expected_leaves = get_expected_order_leaves();
-            let mut test_tree = create_binary_tree_for_testing();
-            // interrupt traversal at all points.
-            for _ in 0..expected_bfs.len() {
-                let mut bfs_iter = test_tree.bfs_iter();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = bfs_iter.next() {
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let bfs_iter_leaves = bfs_iter.leaves();
-                for (i, value) in bfs_iter_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut bfs_iter_mut = test_tree.bfs_iter_mut();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = bfs_iter_mut.next() {
-                    if *value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let bfs_iter_mut_leaves = bfs_iter_mut.leaves();
-                for (i, value) in bfs_iter_mut_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], *value);
-                }
-
-                let mut bfs = test_tree.clone().bfs();
-                let mut num_leaves_seen = 0;
-                while let Some(value) = bfs.next() {
-                    if value == expected_leaves[num_leaves_seen] {
-                        num_leaves_seen += 1;
-                    }
-                }
-
-                let bfs_leaves = bfs.leaves();
-                for (i, value) in bfs_leaves.enumerate() {
-                    assert_eq!(expected_leaves[i + num_leaves_seen], value);
-                }
-            }
-        }
-    }
-
-    mod get_at_path_tests {
-        use alloc::boxed::Box;
-
-        use super::create_binary_tree_for_testing;
-        use super::create_tree_for_testing;
-
-        fn get_tree_path_value_pairs() -> Box<[(Box<[usize]>, Option<usize>)]> {
-            Box::new([
-                (Box::new([]), Some(0)),
-                (Box::new([0]), Some(1)),
-                (Box::new([1]), Some(2)),
-                (Box::new([2]), None),
-                (Box::new([0, 0]), Some(3)),
-                (Box::new([0, 1]), Some(4)),
-                (Box::new([0, 2]), None),
-                (Box::new([1, 0]), Some(5)),
-                (Box::new([1, 1]), Some(6)),
-                (Box::new([1, 2]), None),
-                (Box::new([1, 1, 0]), Some(7)),
-                (Box::new([1, 1, 1]), None),
-                (Box::new([1, 1, 1, 0]), None),
-                (Box::new([1, 1, 0, 0]), Some(8)),
-                (Box::new([1, 1, 0, 1]), None),
-                (Box::new([1, 1, 0, 0, 0]), Some(9)),
-                (Box::new([1, 1, 0, 0, 1]), None),
-                (Box::new([1, 1, 0, 0, 0, 0]), Some(10)),
-                (Box::new([1, 1, 0, 0, 0, 1]), None),
-                (Box::new([1, 1, 0, 0, 0, 0, 0]), None),
-            ])
-        }
-
-        #[test]
-        fn tree_at_path() {
-            use super::OwnedTreeNode;
-
-            let tree = create_tree_for_testing();
-            for path_value_pair in get_tree_path_value_pairs() {
-                assert_eq!(
-                    path_value_pair.1,
-                    tree.clone()
-                        .at_path(&path_value_pair.0)
-                        .map(|tree| tree.value)
-                )
-            }
-        }
-
-        #[test]
-        fn tree_at_path_ref() {
-            use super::BorrowedTreeNode;
-
-            let tree = create_tree_for_testing();
-            for path_value_pair in get_tree_path_value_pairs() {
-                assert_eq!(
-                    path_value_pair.1,
-                    tree.at_path_ref(&path_value_pair.0).map(|tree| tree.value)
-                )
-            }
-        }
-
-        #[test]
-        fn tree_at_path_mut() {
-            use super::MutBorrowedTreeNode;
-
-            let mut tree = create_tree_for_testing();
-            for path_value_pair in get_tree_path_value_pairs() {
-                assert_eq!(
-                    path_value_pair.1,
-                    tree.at_path_mut(&path_value_pair.0).map(|tree| tree.value)
-                )
-            }
-        }
-
-        fn get_binary_tree_path_value_pairs() -> Box<[(Box<[usize]>, Option<usize>)]> {
-            Box::new([
-                (Box::new([]), Some(0)),
-                (Box::new([0]), Some(1)),
-                (Box::new([1]), Some(2)),
-                (Box::new([2]), None),
-                (Box::new([0, 0]), Some(3)),
-                (Box::new([0, 1]), Some(4)),
-                (Box::new([0, 2]), None),
-                (Box::new([1, 0]), Some(5)),
-                (Box::new([1, 1]), Some(6)),
-                (Box::new([1, 2]), None),
-                (Box::new([1, 1, 0]), Some(7)),
-                (Box::new([1, 1, 1]), None),
-                (Box::new([1, 1, 1, 0]), None),
-                (Box::new([1, 1, 0, 1]), Some(8)),
-                (Box::new([1, 1, 0, 0]), None),
-                (Box::new([1, 1, 0, 1, 0]), Some(9)),
-                (Box::new([1, 1, 0, 1, 1]), None),
-                (Box::new([1, 1, 0, 1, 0, 1]), Some(10)),
-                (Box::new([1, 1, 0, 1, 0, 0]), None),
-                (Box::new([1, 1, 0, 1, 0, 1, 0]), None),
-            ])
-        }
-
-        #[test]
-        fn binary_tree_at_path() {
-            use crate::prelude::OwnedBinaryTreeNode;
-
-            let binary_tree = create_binary_tree_for_testing();
-            for path_value_pair in get_binary_tree_path_value_pairs() {
-                assert_eq!(
-                    path_value_pair.1,
-                    binary_tree
-                        .clone()
-                        .at_path(&path_value_pair.0)
-                        .map(|tree| tree.value)
-                )
-            }
-        }
-
-        #[test]
-        fn binary_tree_at_path_ref() {
-            use crate::prelude::BorrowedBinaryTreeNode;
-
-            let binary_tree = create_binary_tree_for_testing();
-            for path_value_pair in get_binary_tree_path_value_pairs() {
-                assert_eq!(
-                    path_value_pair.1,
-                    binary_tree
-                        .at_path_ref(&path_value_pair.0)
-                        .map(|tree| tree.value)
-                )
-            }
-        }
-
-        #[test]
-        fn binary_tree_at_path_mut() {
-            use crate::prelude::MutBorrowedBinaryTreeNode;
-
-            let mut binary_tree = create_binary_tree_for_testing();
-            for path_value_pair in get_binary_tree_path_value_pairs() {
-                assert_eq!(
-                    path_value_pair.1,
-                    binary_tree
-                        .at_path_mut(&path_value_pair.0)
-                        .map(|tree| tree.value)
-                )
-            }
-        }
-    }
-
-    mod prune_tests {
-        use alloc::vec::Vec;
-        use alloc::{boxed::Box, vec};
-
-        use crate::prelude::{
-            tests::{create_binary_tree_for_testing, create_tree_for_testing},
-            BinaryTree, OwnedTreeNode, Tree,
-        };
-        use crate::prelude::{
-            BinaryTreeCollectionIterator, BorrowedBinaryTreeNode,
-            BorrowedIntoIteratorOfBinaryTrees, BorrowedIntoIteratorOfTrees, BorrowedTreeNode,
-            MutBorrowedBinaryTreeNode, MutBorrowedIntoIteratorOfBinaryTrees,
-            MutBorrowedIntoIteratorOfTrees, MutBorrowedTreeNode, OwnedBinaryTreeNode,
-            OwnedIntoIteratorOfBinaryTrees, OwnedIntoIteratorOfTrees, TreeCollectionIterator,
-            TreeCollectionIteratorBase,
-        };
-
-        #[test]
-        fn prune_tree() {
-            let tree = create_tree_for_testing();
-
-            let expected = Some(Tree {
-                value: 0,
-                children: vec![Tree {
-                    value: 1,
-                    children: vec![
-                        Tree {
-                            value: 3,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: 4,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            });
-
-            assert_eq!(None, tree.clone().prune(|_| true));
-            assert_eq!(None, tree.clone().prune(|item| *item == 0));
-            assert_eq!(expected, tree.clone().prune(|item| *item == 2));
-            assert_eq!(
-                expected,
-                tree.clone()
-                    .prune_path(|path, _| matches!(path.get(0), Some(1)))
-            );
-
-            let unevenly_pruned_expected = Some(Tree {
-                value: 0,
-                children: vec![Tree {
-                    value: 1,
-                    children: vec![Tree {
-                        value: 4,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            });
-            let unevenly_pruned_source = tree.prune(|item| *item == 2).unwrap();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.clone().prune(|item| *item == 3)
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .clone()
-                    .prune_path(|path, _| matches!(path.get(1), Some(0)))
-            );
-        }
-
-        #[test]
-        fn prune_tree_iter() {
-            let tree = create_tree_for_testing();
-
-            let expected = Some(Tree {
-                value: &0,
-                children: vec![Tree {
-                    value: &1,
-                    children: vec![
-                        Tree {
-                            value: &3,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &4,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            });
-
-            assert_eq!(None, tree.prune_ref(|_| true));
-            assert_eq!(None, tree.prune_ref(|item| **item == 0));
-            assert_eq!(expected, tree.prune_ref(|item| **item == 2));
-            assert_eq!(
-                expected,
-                tree.prune_path_ref(|path, _| matches!(path.get(0), Some(1)))
-            );
-
-            let unevenly_pruned_expected = Some(Tree {
-                value: &0,
-                children: vec![Tree {
-                    value: &1,
-                    children: vec![Tree {
-                        value: &4,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            });
-            let unevenly_pruned_source = tree.prune_ref(|item| **item == 2).unwrap().cloned();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_ref(|item| **item == 3)
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_path_ref(|path, _| matches!(path.get(1), Some(0)))
-            );
-        }
-
-        #[test]
-        fn prune_tree_mut() {
-            let mut zero = 0;
-            let mut one = 1;
-            let mut three = 3;
-            let mut four = 4;
-
-            let mut tree = create_tree_for_testing();
-
-            let expected = Some(Tree {
-                value: &mut zero,
-                children: vec![Tree {
-                    value: &mut one,
-                    children: vec![
-                        Tree {
-                            value: &mut three,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &mut four,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            });
-
-            assert_eq!(None, tree.prune_mut(|_| true));
-            assert_eq!(None, tree.prune_mut(|item| **item == 0));
-            assert_eq!(expected, tree.prune_mut(|item| **item == 2));
-            assert_eq!(
-                expected,
-                tree.prune_path_mut(|path, _| matches!(path.get(0), Some(1)))
-            );
-
-            let mut unevenly_pruned_source = tree.prune_mut(|item| **item == 2).unwrap().cloned();
-            let unevenly_pruned_expected = Some(Tree {
-                value: &mut zero,
-                children: vec![Tree {
-                    value: &mut one,
-                    children: vec![Tree {
-                        value: &mut four,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            });
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_mut(|item| **item == 3)
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_path_mut(|path, _| matches!(path.get(1), Some(0)))
-            );
-        }
-
-        #[test]
-        fn prune_binary_tree() {
-            let tree = create_binary_tree_for_testing();
-
-            let expected = Some(BinaryTree {
-                value: 0,
-                left: Some(Box::new(BinaryTree {
-                    value: 1,
-                    left: Some(Box::new(BinaryTree {
-                        value: 3,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            });
-
-            assert_eq!(None, tree.clone().prune(|_| true));
-            assert_eq!(None, tree.clone().prune(|item| *item == 0));
-            assert_eq!(expected, tree.clone().prune(|item| *item == 2));
-            assert_eq!(
-                expected,
-                tree.clone()
-                    .prune_path(|path, _| matches!(path.get(0), Some(1)))
-            );
-
-            let unevenly_pruned_expected = Some(BinaryTree {
-                value: 0,
-                left: Some(Box::new(BinaryTree {
-                    value: 1,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: 4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            });
-            let unevenly_pruned_source = tree.prune(|item| *item == 2).unwrap();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.clone().prune(|item| *item == 3)
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .clone()
-                    .prune_path(|path, _| matches!(path.get(1), Some(0)))
-            );
-        }
-
-        #[test]
-        fn prune_binary_tree_iter() {
-            let tree = create_binary_tree_for_testing();
-
-            let expected = Some(BinaryTree {
-                value: &0,
-                left: Some(Box::new(BinaryTree {
-                    value: &1,
-                    left: Some(Box::new(BinaryTree {
-                        value: &3,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            });
-
-            assert_eq!(None, tree.prune_ref(|_| true));
-            assert_eq!(None, tree.prune_ref(|item| **item == 0));
-            assert_eq!(expected, tree.prune_ref(|item| **item == 2));
-            assert_eq!(
-                expected,
-                tree.prune_path_ref(|path, _| matches!(path.get(0), Some(1)))
-            );
-
-            let unevenly_pruned_expected = Some(BinaryTree {
-                value: &0,
-                left: Some(Box::new(BinaryTree {
-                    value: &1,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: &4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            });
-            let unevenly_pruned_source = tree.prune_ref(|item| **item == 2).unwrap().cloned();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_ref(|item| **item == 3)
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_path_ref(|path, _| matches!(path.get(1), Some(0)))
-            );
-        }
-
-        #[test]
-        fn prune_binary_tree_mut() {
-            let mut tree = create_binary_tree_for_testing();
-
-            let mut zero = 0;
-            let mut one = 1;
-            let mut three = 3;
-            let mut four = 4;
-
-            let expected = Some(BinaryTree {
-                value: &mut zero,
-                left: Some(Box::new(BinaryTree {
-                    value: &mut one,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut three,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut four,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            });
-
-            assert_eq!(None, tree.prune_mut(|_| true));
-            assert_eq!(None, tree.prune_mut(|item| **item == 0));
-            assert_eq!(expected, tree.prune_mut(|item| **item == 2));
-            assert_eq!(
-                expected,
-                tree.prune_path_mut(|path, _| matches!(path.get(0), Some(1)))
-            );
-
-            let mut unevenly_pruned_source = tree.prune_mut(|item| **item == 2).unwrap().cloned();
-            let unevenly_pruned_expected = Some(BinaryTree {
-                value: &mut zero,
-                left: Some(Box::new(BinaryTree {
-                    value: &mut one,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut four,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            });
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_mut(|item| **item == 3)
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source.prune_path_mut(|path, _| matches!(path.get(1), Some(0)))
-            );
-        }
-
-        #[test]
-        fn prune_tree_collection() {
-            let trees = vec![create_tree_for_testing(), create_tree_for_testing()];
-
-            let expected_tree = Tree {
-                value: 0,
-                children: vec![Tree {
-                    value: 1,
-                    children: vec![
-                        Tree {
-                            value: 3,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: 4,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            };
-
-            let expected: Vec<Tree<usize>> = vec![expected_tree.clone(), expected_tree];
-
-            assert_eq!(
-                Vec::<Tree<usize>>::with_capacity(0),
-                trees.clone().prune_each(|_| true).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                Vec::<Tree<usize>>::with_capacity(0),
-                trees
-                    .clone()
-                    .prune_each(|item| *item == 0)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .clone()
-                    .prune_each(|item| *item == 2)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .clone()
-                    .prune_path_each(|path, _| matches!(path.get(1), Some(1)))
-                    .collect::<Vec<_>>()
-            );
-
-            let unevenly_pruned_expected_tree = Tree {
-                value: 0,
-                children: vec![Tree {
-                    value: 1,
-                    children: vec![Tree {
-                        value: 4,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            };
-
-            let unevenly_pruned_expected = vec![
-                unevenly_pruned_expected_tree.clone(),
-                unevenly_pruned_expected_tree,
-            ];
-            let unevenly_pruned_source = trees.prune_each(|item| *item == 2).collect::<Vec<_>>();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .clone()
-                    .prune_each(|item| *item == 3)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .clone()
-                    .prune_path_each(|path, _| matches!(path.get(2), Some(0)))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn prune_tree_collection_ref() {
-            let trees = vec![create_tree_for_testing(), create_tree_for_testing()];
-
-            let expected_tree = Tree {
-                value: &0,
-                children: vec![Tree {
-                    value: &1,
-                    children: vec![
-                        Tree {
-                            value: &3,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &4,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            };
-
-            let expected: Vec<Tree<&usize>> = vec![expected_tree.clone(), expected_tree];
-
-            assert_eq!(
-                Vec::<Tree<&usize>>::with_capacity(0),
-                trees.prune_each_ref(|_| true).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                Vec::<Tree<&usize>>::with_capacity(0),
-                trees.prune_each_ref(|item| **item == 0).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees.prune_each_ref(|item| **item == 2).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .prune_path_each_ref(|path, _| matches!(path.get(1), Some(1)))
-                    .collect::<Vec<_>>()
-            );
-
-            let unevenly_pruned_expected_tree = Tree {
-                value: &0,
-                children: vec![Tree {
-                    value: &1,
-                    children: vec![Tree {
-                        value: &4,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            };
-
-            let unevenly_pruned_source = trees.prune_each(|item| *item == 2).collect::<Vec<_>>();
-            let unevenly_pruned_expected = vec![
-                unevenly_pruned_expected_tree.clone(),
-                unevenly_pruned_expected_tree,
-            ];
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_each_ref(|item| **item == 3)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_path_each_ref(|path, _| matches!(path.get(2), Some(0)))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn prune_tree_collection_mut() {
-            let mut trees = vec![create_tree_for_testing(), create_tree_for_testing()];
-
-            let mut zero = 0;
-            let mut one = 1;
-            let mut three = 3;
-            let mut four = 4;
-
-            let mut second_zero = 0;
-            let mut second_one = 1;
-            let mut second_three = 3;
-            let mut second_four = 4;
-
-            let expected_tree = Tree {
-                value: &mut zero,
-                children: vec![Tree {
-                    value: &mut one,
-                    children: vec![
-                        Tree {
-                            value: &mut three,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &mut four,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            };
-            let expected_tree2 = Tree {
-                value: &mut second_zero,
-                children: vec![Tree {
-                    value: &mut second_one,
-                    children: vec![
-                        Tree {
-                            value: &mut second_three,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &mut second_four,
-                            children: Vec::with_capacity(0),
-                        },
-                    ],
-                }],
-            };
-
-            let expected: Vec<Tree<&mut usize>> = vec![expected_tree, expected_tree2];
-
-            assert_eq!(
-                Vec::<Tree<&mut usize>>::with_capacity(0),
-                trees.prune_each_mut(|_| true).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                Vec::<Tree<&mut usize>>::with_capacity(0),
-                trees.prune_each_mut(|item| **item == 0).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees.prune_each_mut(|item| **item == 2).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .prune_path_each_mut(|path, _| matches!(path.get(1), Some(1)))
-                    .collect::<Vec<_>>()
-            );
-
-            let unevenly_pruned_expected_tree = Tree {
-                value: &mut zero,
-                children: vec![Tree {
-                    value: &mut one,
-                    children: vec![Tree {
-                        value: &mut four,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            };
-
-            let unevenly_pruned_expected_tree2 = Tree {
-                value: &mut second_zero,
-                children: vec![Tree {
-                    value: &mut second_one,
-                    children: vec![Tree {
-                        value: &mut second_four,
-                        children: Vec::with_capacity(0),
-                    }],
-                }],
-            };
-
-            let mut unevenly_pruned_source = trees
-                .into_pipeline_mut()
-                .prune(|item| **item == 2)
-                .map_trees(|item| item.clone())
-                .trees()
-                .collect::<Vec<_>>();
-
-            let unevenly_pruned_expected = vec![
-                unevenly_pruned_expected_tree,
-                unevenly_pruned_expected_tree2,
-            ];
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_each_mut(|item| **item == 3)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_path_each_mut(|path, _| matches!(path.get(2), Some(0)))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn prune_binary_tree_collection() {
-            let trees = vec![
-                create_binary_tree_for_testing(),
-                create_binary_tree_for_testing(),
-            ];
-
-            let expected_tree = BinaryTree {
-                value: 0,
-                left: Some(Box::new(BinaryTree {
-                    value: 1,
-                    left: Some(Box::new(BinaryTree {
-                        value: 3,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let expected: Vec<BinaryTree<usize>> = vec![expected_tree.clone(), expected_tree];
-
-            assert_eq!(
-                Vec::<BinaryTree<usize>>::with_capacity(0),
-                trees.clone().prune_each(|_| true).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                Vec::<BinaryTree<usize>>::with_capacity(0),
-                trees
-                    .clone()
-                    .prune_each(|item| *item == 0)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .clone()
-                    .prune_each(|item| *item == 2)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .clone()
-                    .prune_path_each(|path, _| matches!(path.get(1), Some(1)))
-                    .collect::<Vec<_>>()
-            );
-
-            let unevenly_pruned_expected_tree = BinaryTree {
-                value: 0,
-                left: Some(Box::new(BinaryTree {
-                    value: 1,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: 4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let unevenly_pruned_expected = vec![
-                unevenly_pruned_expected_tree.clone(),
-                unevenly_pruned_expected_tree,
-            ];
-            let unevenly_pruned_source = trees.prune_each(|item| *item == 2).collect::<Vec<_>>();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .clone()
-                    .prune_each(|item| *item == 3)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .clone()
-                    .prune_path_each(|path, _| matches!(path.get(2), Some(0)))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn prune_binary_tree_collection_ref() {
-            let trees = vec![
-                create_binary_tree_for_testing(),
-                create_binary_tree_for_testing(),
-            ];
-
-            let expected_tree = BinaryTree {
-                value: &0,
-                left: Some(Box::new(BinaryTree {
-                    value: &1,
-                    left: Some(Box::new(BinaryTree {
-                        value: &3,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let expected: Vec<BinaryTree<&usize>> = vec![expected_tree.clone(), expected_tree];
-
-            assert_eq!(
-                Vec::<BinaryTree<&usize>>::with_capacity(0),
-                trees.prune_each_ref(|_| true).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                Vec::<BinaryTree<&usize>>::with_capacity(0),
-                trees.prune_each_ref(|item| **item == 0).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees.prune_each_ref(|item| **item == 2).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .prune_path_each_ref(|path, _| matches!(path.get(1), Some(1)))
-                    .collect::<Vec<_>>()
-            );
-
-            let unevenly_pruned_expected_tree = BinaryTree {
-                value: &0,
-                left: Some(Box::new(BinaryTree {
-                    value: &1,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: &4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let unevenly_pruned_expected = vec![
-                unevenly_pruned_expected_tree.clone(),
-                unevenly_pruned_expected_tree,
-            ];
-            let unevenly_pruned_source = trees.prune_each(|item| *item == 2).collect::<Vec<_>>();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_each_ref(|item| **item == 3)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_path_each_ref(|path, _| matches!(path.get(2), Some(0)))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn prune_binary_tree_collection_mut() {
-            let mut trees = vec![
-                create_binary_tree_for_testing(),
-                create_binary_tree_for_testing(),
-            ];
-
-            let mut zero = 0;
-            let mut one = 1;
-            let mut three = 3;
-            let mut four = 4;
-
-            let mut second_zero = 0;
-            let mut second_one = 1;
-            let mut second_three = 3;
-            let mut second_four = 4;
-
-            let expected_tree = BinaryTree {
-                value: &mut zero,
-                left: Some(Box::new(BinaryTree {
-                    value: &mut one,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut three,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut four,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let expected_tree2 = BinaryTree {
-                value: &mut second_zero,
-                left: Some(Box::new(BinaryTree {
-                    value: &mut second_one,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut second_three,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut second_four,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let expected: Vec<BinaryTree<&mut usize>> = vec![expected_tree, expected_tree2];
-
-            assert_eq!(
-                Vec::<BinaryTree<&mut usize>>::with_capacity(0),
-                trees.prune_each_mut(|_| true).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                Vec::<BinaryTree<&mut usize>>::with_capacity(0),
-                trees.prune_each_mut(|item| **item == 0).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees.prune_each_mut(|item| **item == 2).collect::<Vec<_>>()
-            );
-            assert_eq!(
-                expected,
-                trees
-                    .prune_path_each_mut(|path, _| matches!(path.get(1), Some(1)))
-                    .collect::<Vec<_>>()
-            );
-
-            let unevenly_pruned_expected_tree = BinaryTree {
-                value: &mut zero,
-                left: Some(Box::new(BinaryTree {
-                    value: &mut one,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut four,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let unevenly_pruned_expected_tree2 = BinaryTree {
-                value: &mut second_zero,
-                left: Some(Box::new(BinaryTree {
-                    value: &mut second_one,
-                    left: None,
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut second_four,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: None,
-            };
-
-            let unevenly_pruned_expected = vec![
-                unevenly_pruned_expected_tree,
-                unevenly_pruned_expected_tree2,
-            ];
-            let mut unevenly_pruned_source = trees
-                .into_pipeline_mut()
-                .prune(|item| **item == 2)
-                .map_trees(|item| item.clone())
-                .trees()
-                .collect::<Vec<_>>();
-
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_each_mut(|item| **item == 3)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                unevenly_pruned_expected,
-                unevenly_pruned_source
-                    .prune_path_each_mut(|path, _| matches!(path.get(2), Some(0)))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn prune_depth_tree() {
-            let tree = create_tree_for_testing();
-            let tree_ref = tree.clone();
-            let mut tree_mut = tree.clone();
-
-            // depth 0
-            assert_eq!(
-                tree.clone().prune_depth(0),
-                Tree {
-                    value: 0,
-                    children: Vec::with_capacity(0)
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(0),
-                Tree {
-                    value: &0,
-                    children: Vec::with_capacity(0)
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(0),
-                Tree {
-                    value: &mut 0,
-                    children: Vec::with_capacity(0),
-                }
-            );
-
-            // depth 1
-            assert_eq!(
-                tree.clone().prune_depth(1),
-                Tree {
-                    value: 0,
-                    children: vec![
-                        Tree {
-                            value: 1,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: 2,
-                            children: Vec::with_capacity(0),
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(1),
-                Tree {
-                    value: &0,
-                    children: vec![
-                        Tree {
-                            value: &1,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &2,
-                            children: Vec::with_capacity(0),
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(1),
-                Tree {
-                    value: &mut 0,
-                    children: vec![
-                        Tree {
-                            value: &mut 1,
-                            children: Vec::with_capacity(0),
-                        },
-                        Tree {
-                            value: &mut 2,
-                            children: Vec::with_capacity(0),
-                        }
-                    ]
-                }
-            );
-
-            // depth 2
-            assert_eq!(
-                tree.clone().prune_depth(2),
-                Tree {
-                    value: 0,
-                    children: vec![
-                        Tree {
-                            value: 1,
-                            children: vec![
-                                Tree {
-                                    value: 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: 2,
-                            children: vec![
-                                Tree {
-                                    value: 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 6,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(2),
-                Tree {
-                    value: &0,
-                    children: vec![
-                        Tree {
-                            value: &1,
-                            children: vec![
-                                Tree {
-                                    value: &3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &2,
-                            children: vec![
-                                Tree {
-                                    value: &5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &6,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(2),
-                Tree {
-                    value: &mut 0,
-                    children: vec![
-                        Tree {
-                            value: &mut 1,
-                            children: vec![
-                                Tree {
-                                    value: &mut 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &mut 2,
-                            children: vec![
-                                Tree {
-                                    value: &mut 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 6,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-
-            // depth 3
-            assert_eq!(
-                tree.clone().prune_depth(3),
-                Tree {
-                    value: 0,
-                    children: vec![
-                        Tree {
-                            value: 1,
-                            children: vec![
-                                Tree {
-                                    value: 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: 2,
-                            children: vec![
-                                Tree {
-                                    value: 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 6,
-                                    children: vec![Tree {
-                                        value: 7,
-                                        children: Vec::with_capacity(0)
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(3),
-                Tree {
-                    value: &0,
-                    children: vec![
-                        Tree {
-                            value: &1,
-                            children: vec![
-                                Tree {
-                                    value: &3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &2,
-                            children: vec![
-                                Tree {
-                                    value: &5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &6,
-                                    children: vec![Tree {
-                                        value: &7,
-                                        children: Vec::with_capacity(0)
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(3),
-                Tree {
-                    value: &mut 0,
-                    children: vec![
-                        Tree {
-                            value: &mut 1,
-                            children: vec![
-                                Tree {
-                                    value: &mut 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &mut 2,
-                            children: vec![
-                                Tree {
-                                    value: &mut 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 6,
-                                    children: vec![Tree {
-                                        value: &mut 7,
-                                        children: Vec::with_capacity(0)
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-
-            // depth 4
-            assert_eq!(
-                tree.clone().prune_depth(4),
-                Tree {
-                    value: 0,
-                    children: vec![
-                        Tree {
-                            value: 1,
-                            children: vec![
-                                Tree {
-                                    value: 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: 2,
-                            children: vec![
-                                Tree {
-                                    value: 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 6,
-                                    children: vec![Tree {
-                                        value: 7,
-                                        children: vec![Tree {
-                                            value: 8,
-                                            children: Vec::with_capacity(0),
-                                        }]
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(4),
-                Tree {
-                    value: &0,
-                    children: vec![
-                        Tree {
-                            value: &1,
-                            children: vec![
-                                Tree {
-                                    value: &3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &2,
-                            children: vec![
-                                Tree {
-                                    value: &5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &6,
-                                    children: vec![Tree {
-                                        value: &7,
-                                        children: vec![Tree {
-                                            value: &8,
-                                            children: Vec::with_capacity(0),
-                                        }]
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(4),
-                Tree {
-                    value: &mut 0,
-                    children: vec![
-                        Tree {
-                            value: &mut 1,
-                            children: vec![
-                                Tree {
-                                    value: &mut 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &mut 2,
-                            children: vec![
-                                Tree {
-                                    value: &mut 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 6,
-                                    children: vec![Tree {
-                                        value: &mut 7,
-                                        children: vec![Tree {
-                                            value: &mut 8,
-                                            children: Vec::with_capacity(0),
-                                        }]
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-
-            // depth 5
-            assert_eq!(
-                tree.clone().prune_depth(5),
-                Tree {
-                    value: 0,
-                    children: vec![
-                        Tree {
-                            value: 1,
-                            children: vec![
-                                Tree {
-                                    value: 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: 2,
-                            children: vec![
-                                Tree {
-                                    value: 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: 6,
-                                    children: vec![Tree {
-                                        value: 7,
-                                        children: vec![Tree {
-                                            value: 8,
-                                            children: vec![Tree {
-                                                value: 9,
-                                                children: Vec::with_capacity(0)
-                                            }],
-                                        }]
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(5),
-                Tree {
-                    value: &0,
-                    children: vec![
-                        Tree {
-                            value: &1,
-                            children: vec![
-                                Tree {
-                                    value: &3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &2,
-                            children: vec![
-                                Tree {
-                                    value: &5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &6,
-                                    children: vec![Tree {
-                                        value: &7,
-                                        children: vec![Tree {
-                                            value: &8,
-                                            children: vec![Tree {
-                                                value: &9,
-                                                children: Vec::with_capacity(0)
-                                            }],
-                                        }]
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(5),
-                Tree {
-                    value: &mut 0,
-                    children: vec![
-                        Tree {
-                            value: &mut 1,
-                            children: vec![
-                                Tree {
-                                    value: &mut 3,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 4,
-                                    children: Vec::with_capacity(0),
-                                }
-                            ],
-                        },
-                        Tree {
-                            value: &mut 2,
-                            children: vec![
-                                Tree {
-                                    value: &mut 5,
-                                    children: Vec::with_capacity(0),
-                                },
-                                Tree {
-                                    value: &mut 6,
-                                    children: vec![Tree {
-                                        value: &mut 7,
-                                        children: vec![Tree {
-                                            value: &mut 8,
-                                            children: vec![Tree {
-                                                value: &mut 9,
-                                                children: Vec::with_capacity(0)
-                                            }],
-                                        }]
-                                    }],
-                                }
-                            ],
-                        }
-                    ]
-                }
-            );
-
-            for depth in 6..10 {
-                assert_eq!(tree.clone().prune_depth(depth), tree);
-                assert_eq!(
-                    tree_ref.map_ref(|item| item),
-                    tree_ref.prune_depth_ref(depth)
-                );
-                assert_eq!(
-                    tree_mut.clone().map_mut(|item| item),
-                    tree_mut.prune_depth_mut(depth)
-                );
-            }
-        }
-
-        #[test]
-        fn prune_depth_binary_tree() {
-            let tree = create_binary_tree_for_testing();
-            let tree_ref = tree.clone();
-            let mut tree_mut = tree.clone();
-
-            // depth 0
-            assert_eq!(
-                tree.clone().prune_depth(0),
-                BinaryTree {
-                    value: 0,
-                    left: None,
-                    right: None,
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(0),
-                BinaryTree {
-                    value: &0,
-                    left: None,
-                    right: None,
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(0),
-                BinaryTree {
-                    value: &mut 0,
-                    left: None,
-                    right: None,
-                }
-            );
-
-            // depth 1
-            assert_eq!(
-                tree.clone().prune_depth(1),
-                BinaryTree {
-                    value: 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: 1,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 2,
-                        left: None,
-                        right: None,
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(1),
-                BinaryTree {
-                    value: &0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &1,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &2,
-                        left: None,
-                        right: None,
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(1),
-                BinaryTree {
-                    value: &mut 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut 1,
-                        left: None,
-                        right: None,
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut 2,
-                        left: None,
-                        right: None,
-                    }))
-                }
-            );
-
-            // depth 2
-            assert_eq!(
-                tree.clone().prune_depth(2),
-                BinaryTree {
-                    value: 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 6,
-                            left: None,
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(2),
-                BinaryTree {
-                    value: &0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &6,
-                            left: None,
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(2),
-                BinaryTree {
-                    value: &mut 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 6,
-                            left: None,
-                            right: None
-                        })),
-                    }))
-                }
-            );
-
-            // depth 3
-            assert_eq!(
-                tree.clone().prune_depth(3),
-                BinaryTree {
-                    value: 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 6,
-                            left: Some(Box::new(BinaryTree {
-                                value: 7,
-                                left: None,
-                                right: None
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(3),
-                BinaryTree {
-                    value: &0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &6,
-                            left: Some(Box::new(BinaryTree {
-                                value: &7,
-                                left: None,
-                                right: None
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(3),
-                BinaryTree {
-                    value: &mut 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 6,
-                            left: Some(Box::new(BinaryTree {
-                                value: &mut 7,
-                                left: None,
-                                right: None
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-
-            // depth 4
-            assert_eq!(
-                tree.clone().prune_depth(4),
-                BinaryTree {
-                    value: 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 6,
-                            left: Some(Box::new(BinaryTree {
-                                value: 7,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: 8,
-                                    left: None,
-                                    right: None
-                                }))
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(4),
-                BinaryTree {
-                    value: &0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &6,
-                            left: Some(Box::new(BinaryTree {
-                                value: &7,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: &8,
-                                    left: None,
-                                    right: None
-                                }))
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(4),
-                BinaryTree {
-                    value: &mut 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 6,
-                            left: Some(Box::new(BinaryTree {
-                                value: &mut 7,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: &mut 8,
-                                    left: None,
-                                    right: None
-                                }))
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-
-            // depth 5
-            assert_eq!(
-                tree.clone().prune_depth(5),
-                BinaryTree {
-                    value: 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: 6,
-                            left: Some(Box::new(BinaryTree {
-                                value: 7,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: 8,
-                                    left: Some(Box::new(BinaryTree {
-                                        value: 9,
-                                        left: None,
-                                        right: None
-                                    })),
-                                    right: None
-                                }))
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_ref.prune_depth_ref(5),
-                BinaryTree {
-                    value: &0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &6,
-                            left: Some(Box::new(BinaryTree {
-                                value: &7,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: &8,
-                                    left: Some(Box::new(BinaryTree {
-                                        value: &9,
-                                        left: None,
-                                        right: None
-                                    })),
-                                    right: None
-                                }))
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-            assert_eq!(
-                tree_mut.prune_depth_mut(5),
-                BinaryTree {
-                    value: &mut 0,
-                    left: Some(Box::new(BinaryTree {
-                        value: &mut 1,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 3,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 4,
-                            left: None,
-                            right: None
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: &mut 2,
-                        left: Some(Box::new(BinaryTree {
-                            value: &mut 5,
-                            left: None,
-                            right: None
-                        })),
-                        right: Some(Box::new(BinaryTree {
-                            value: &mut 6,
-                            left: Some(Box::new(BinaryTree {
-                                value: &mut 7,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: &mut 8,
-                                    left: Some(Box::new(BinaryTree {
-                                        value: &mut 9,
-                                        left: None,
-                                        right: None
-                                    })),
-                                    right: None
-                                }))
-                            })),
-                            right: None
-                        })),
-                    }))
-                }
-            );
-
-            for depth in 6..10 {
-                assert_eq!(tree.clone().prune_depth(depth), tree);
-                assert_eq!(
-                    tree_ref.prune_depth_ref(depth),
-                    tree_ref.map_ref(|item| item)
-                );
-                assert_eq!(
-                    tree.clone().prune_depth_mut(depth),
-                    tree_mut.map_mut(|item| item)
-                );
-            }
-        }
-    }
-
-    mod map_tests {
-        use alloc::boxed::Box;
-        use alloc::vec::Vec;
-        use alloc::{string::ToString, vec};
-
-        use crate::prelude::{
-            BinaryTree, BorrowedBinaryTreeNode, BorrowedIntoIteratorOfBinaryTrees,
-            BorrowedIntoIteratorOfTrees, BorrowedTreeNode, MutBorrowedBinaryTreeNode,
-            MutBorrowedIntoIteratorOfBinaryTrees, MutBorrowedIntoIteratorOfTrees,
-            MutBorrowedTreeNode, OwnedBinaryTreeNode, OwnedIntoIteratorOfBinaryTrees,
-            OwnedIntoIteratorOfTrees, OwnedTreeNode, Tree,
-        };
-
-        #[test]
-        fn map_tree() {
-            let mut original = Tree {
-                value: "0_0".to_string(),
-                children: vec![
-                    Tree {
-                        value: "1_1".to_string(),
-                        children: vec![
-                            Tree {
-                                value: "3_3".to_string(),
-                                children: vec![Tree {
-                                    value: "5_5".to_string(),
-                                    children: Vec::with_capacity(0),
-                                }],
-                            },
-                            Tree {
-                                value: "4_4".to_string(),
-                                children: Vec::with_capacity(0),
-                            },
-                        ],
-                    },
-                    Tree {
-                        value: "2_2".to_string(),
-                        children: Vec::with_capacity(0),
-                    },
-                ],
-            };
-
-            let expected = Tree {
-                value: 0,
-                children: vec![
-                    Tree {
-                        value: 1,
-                        children: vec![
-                            Tree {
-                                value: 3,
-                                children: vec![Tree {
-                                    value: 5,
-                                    children: Vec::with_capacity(0),
-                                }],
-                            },
-                            Tree {
-                                value: 4,
-                                children: Vec::with_capacity(0),
-                            },
-                        ],
-                    },
-                    Tree {
-                        value: 2,
-                        children: Vec::with_capacity(0),
-                    },
-                ],
-            };
-
-            assert_eq!(
-                expected,
-                original.map_ref(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-            );
-            assert_eq!(
-                expected,
-                original.map_mut(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-            );
-            assert_eq!(
-                expected,
-                original.map(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-            );
-        }
-
-        #[test]
-        fn map_binary_tree() {
-            let mut original = BinaryTree {
-                value: "0_0".to_string(),
-                left: Some(Box::new(BinaryTree {
-                    value: "1_1".to_string(),
-                    left: Some(Box::new(BinaryTree {
-                        value: "3_3".to_string(),
-                        left: None,
-                        right: Some(Box::new(BinaryTree {
-                            value: "5_5".to_string(),
-                            left: None,
-                            right: None,
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: "4_4".to_string(),
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: Some(Box::new(BinaryTree {
-                    value: "2_2".to_string(),
-                    left: None,
-                    right: None,
-                })),
-            };
-
-            let expected = BinaryTree {
-                value: 0,
-                left: Some(Box::new(BinaryTree {
-                    value: 1,
-                    left: Some(Box::new(BinaryTree {
-                        value: 3,
-                        left: None,
-                        right: Some(Box::new(BinaryTree {
-                            value: 5,
-                            left: None,
-                            right: None,
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: Some(Box::new(BinaryTree {
-                    value: 2,
-                    left: None,
-                    right: None,
-                })),
-            };
-
-            assert_eq!(
-                expected,
-                original.map_ref(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-            );
-            assert_eq!(
-                expected,
-                original.map_mut(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-            );
-            assert_eq!(
-                expected,
-                original.map(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-            );
-        }
-
-        #[test]
-        fn map_tree_collection() {
-            let original = Tree {
-                value: "0_0".to_string(),
-                children: vec![
-                    Tree {
-                        value: "1_1".to_string(),
-                        children: vec![
-                            Tree {
-                                value: "3_3".to_string(),
-                                children: vec![Tree {
-                                    value: "5_5".to_string(),
-                                    children: Vec::with_capacity(0),
-                                }],
-                            },
-                            Tree {
-                                value: "4_4".to_string(),
-                                children: Vec::with_capacity(0),
-                            },
-                        ],
-                    },
-                    Tree {
-                        value: "2_2".to_string(),
-                        children: Vec::with_capacity(0),
-                    },
-                ],
-            };
-
-            let mut original_collection = vec![original.clone(), original];
-
-            let expected_result = Tree {
-                value: 0,
-                children: vec![
-                    Tree {
-                        value: 1,
-                        children: vec![
-                            Tree {
-                                value: 3,
-                                children: vec![Tree {
-                                    value: 5,
-                                    children: Vec::with_capacity(0),
-                                }],
-                            },
-                            Tree {
-                                value: 4,
-                                children: Vec::with_capacity(0),
-                            },
-                        ],
-                    },
-                    Tree {
-                        value: 2,
-                        children: Vec::with_capacity(0),
-                    },
-                ],
-            };
-
-            assert_eq!(
-                vec![expected_result.clone(), expected_result.clone()],
-                original_collection
-                    .map_each_ref(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![expected_result.clone(), expected_result.clone()],
-                original_collection
-                    .map_each_mut(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![expected_result.clone(), expected_result],
-                original_collection
-                    .map_each(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn map_binary_tree_collection() {
-            let original = BinaryTree {
-                value: "0_0".to_string(),
-                left: Some(Box::new(BinaryTree {
-                    value: "1_1".to_string(),
-                    left: Some(Box::new(BinaryTree {
-                        value: "3_3".to_string(),
-                        left: None,
-                        right: Some(Box::new(BinaryTree {
-                            value: "5_5".to_string(),
-                            left: None,
-                            right: None,
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: "4_4".to_string(),
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: Some(Box::new(BinaryTree {
-                    value: "2_2".to_string(),
-                    left: None,
-                    right: None,
-                })),
-            };
-
-            let mut original_collection = vec![original.clone(), original];
-
-            let expected_result = BinaryTree {
-                value: 0,
-                left: Some(Box::new(BinaryTree {
-                    value: 1,
-                    left: Some(Box::new(BinaryTree {
-                        value: 3,
-                        left: None,
-                        right: Some(Box::new(BinaryTree {
-                            value: 5,
-                            left: None,
-                            right: None,
-                        })),
-                    })),
-                    right: Some(Box::new(BinaryTree {
-                        value: 4,
-                        left: None,
-                        right: None,
-                    })),
-                })),
-                right: Some(Box::new(BinaryTree {
-                    value: 2,
-                    left: None,
-                    right: None,
-                })),
-            };
-
-            assert_eq!(
-                vec![expected_result.clone(), expected_result.clone()],
-                original_collection
-                    .map_each_ref(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![expected_result.clone(), expected_result.clone()],
-                original_collection
-                    .map_each_mut(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![expected_result.clone(), expected_result],
-                original_collection
-                    .map_each(|value| value.split('_').next().unwrap().parse::<i32>().unwrap())
-                    .collect::<Vec<_>>()
-            );
-        }
-    }
-
-    mod fold_tests {
-        use alloc::vec;
-        use alloc::vec::Vec;
-
-        use crate::prelude::{
-            tests::{create_binary_tree_for_testing, create_tree_for_testing},
-            BorrowedBinaryTreeNode, BorrowedIntoIteratorOfBinaryTrees, BorrowedIntoIteratorOfTrees,
-            BorrowedTreeNode, MutBorrowedBinaryTreeNode, MutBorrowedIntoIteratorOfBinaryTrees,
-            MutBorrowedIntoIteratorOfTrees, MutBorrowedTreeNode, OwnedBinaryTreeNode,
-            OwnedIntoIteratorOfBinaryTrees, OwnedIntoIteratorOfTrees, OwnedTreeNode,
-        };
-
-        #[test]
-        fn fold_tree() {
-            let mut tree = create_tree_for_testing();
-            assert_eq!(
-                55,
-                tree.fold_ref(|child_accs, value| child_accs.into_iter().sum::<usize>() + *value)
-            );
-            assert_eq!(
-                55,
-                tree.fold_mut(|child_accs, value| child_accs.into_iter().sum::<usize>() + *value)
-            );
-            assert_eq!(
-                55,
-                tree.fold(|child_accs, value| child_accs.into_iter().sum::<usize>() + value)
-            );
-        }
-
-        #[test]
-        fn fold_binary_tree() {
-            let mut binary_tree = create_binary_tree_for_testing();
-            assert_eq!(
-                55,
-                binary_tree.fold_ref(|child_accs, value| child_accs
-                    .into_iter()
-                    .flat_map(|opt| opt)
-                    .sum::<usize>()
-                    + *value)
-            );
-            assert_eq!(
-                55,
-                binary_tree.fold_mut(|child_accs, value| child_accs
-                    .into_iter()
-                    .flat_map(|opt| opt)
-                    .sum::<usize>()
-                    + *value)
-            );
-            assert_eq!(
-                55,
-                binary_tree.fold(|child_accs, value| child_accs
-                    .into_iter()
-                    .flat_map(|opt| opt)
-                    .sum::<usize>()
-                    + value)
-            );
-        }
-
-        #[test]
-        fn fold_trees() {
-            let mut trees = vec![
-                create_tree_for_testing(),
-                create_tree_for_testing(),
-                create_tree_for_testing(),
-            ];
-
-            assert_eq!(
-                vec![55, 55, 55],
-                trees
-                    .fold_each_ref(
-                        |children_accs, value| children_accs.into_iter().sum::<usize>() + *value
-                    )
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![55, 55, 55],
-                trees
-                    .fold_each_mut(
-                        |children_accs, value| children_accs.into_iter().sum::<usize>() + *value
-                    )
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![55, 55, 55],
-                trees
-                    .fold_each(
-                        |children_accs, value| children_accs.into_iter().sum::<usize>() + value
-                    )
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        #[test]
-        fn fold_binary_trees() {
-            let mut trees = vec![
-                create_binary_tree_for_testing(),
-                create_binary_tree_for_testing(),
-                create_binary_tree_for_testing(),
-            ];
-
-            assert_eq!(
-                vec![55, 55, 55],
-                trees
-                    .fold_each_ref(|child_accs, value| child_accs
-                        .into_iter()
-                        .flat_map(|opt| opt)
-                        .sum::<usize>()
-                        + *value)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![55, 55, 55],
-                trees
-                    .fold_each_mut(|child_accs, value| child_accs
-                        .into_iter()
-                        .flat_map(|opt| opt)
-                        .sum::<usize>()
-                        + *value)
-                    .collect::<Vec<_>>()
-            );
-            assert_eq!(
-                vec![55, 55, 55],
-                trees
-                    .fold_each(|child_accs, value| child_accs
-                        .into_iter()
-                        .flat_map(|opt| opt)
-                        .sum::<usize>()
-                        + value)
-                    .collect::<Vec<_>>()
-            );
-        }
-    }
-
-    macro_rules! assert_len {
-        ($expected: expr, $iter: expr) => {
-            assert_len!($expected, $iter, "");
-        };
-        ($expected: expr, $iter: expr, $message: expr) => {
-            let mut count = 0;
-            $iter.for_each(|_| count += 1);
-            assert_eq!($expected, count, "{}", $message);
-        };
-    }
-    use assert_len;
-
-    fn get_expected_metadata_for_value(val: usize) -> &'static [usize] {
-        match val {
-            0 => &[0],
-            1 => &[0, 1],
-            2 => &[0, 2],
-            3 => &[0, 1, 3],
-            4 => &[0, 1, 4],
-            5 => &[0, 2, 5],
-            6 => &[0, 2, 6],
-            7 => &[0, 2, 6, 7],
-            8 => &[0, 2, 6, 7, 8],
-            9 => &[0, 2, 6, 7, 8, 9],
-            10 => &[0, 2, 6, 7, 8, 9, 10],
-            _ => panic!("unexpected value"),
-        }
-    }
-
-    fn create_trees_for_testing() -> Vec<Tree<usize>> {
-        vec![create_tree_for_testing()]
-    }
-
-    pub(crate) fn create_tree_for_testing() -> Tree<usize> {
-        Tree {
-            value: 0,
-            children: vec![
-                Tree {
-                    value: 1,
-                    children: vec![
-                        Tree {
-                            value: 3,
-                            children: Vec::new(),
-                        },
-                        Tree {
-                            value: 4,
-                            children: Vec::new(),
-                        },
-                    ],
-                },
-                Tree {
-                    value: 2,
-                    children: vec![
-                        Tree {
-                            value: 5,
-                            children: Vec::new(),
-                        },
-                        Tree {
-                            value: 6,
-                            children: vec![Tree {
-                                value: 7,
-                                children: vec![Tree {
-                                    value: 8,
-                                    children: vec![Tree {
-                                        value: 9,
-                                        children: vec![Tree {
-                                            value: 10,
-                                            children: Vec::new(),
-                                        }],
-                                    }],
-                                }],
-                            }],
-                        },
-                    ],
-                },
-            ],
-        }
-    }
-
-    pub fn create_binary_tree_for_testing() -> BinaryTree<usize> {
-        BinaryTree {
-            value: 0,
-            left: Some(Box::new(BinaryTree {
-                value: 1,
-                left: Some(Box::new(BinaryTree {
-                    value: 3,
-                    left: None,
-                    right: None,
-                })),
-                right: Some(Box::new(BinaryTree {
-                    value: 4,
-                    left: None,
-                    right: None,
-                })),
-            })),
-            right: Some(Box::new(BinaryTree {
-                value: 2,
-                left: Some(Box::new(BinaryTree {
-                    value: 5,
-                    left: None,
-                    right: None,
-                })),
-                right: Some(Box::new(BinaryTree {
-                    value: 6,
-                    left: Some(Box::new(BinaryTree {
-                        value: 7,
-                        left: None,
-                        right: Some(Box::new(BinaryTree {
-                            value: 8,
-                            left: Some(Box::new(BinaryTree {
-                                value: 9,
-                                left: None,
-                                right: Some(Box::new(BinaryTree {
-                                    value: 10,
-                                    left: None,
-                                    right: None,
-                                })),
-                            })),
-                            right: None,
-                        })),
-                    })),
-                    right: None,
-                })),
-            })),
-        }
-    }
-}
